refactor: move DomRootChecker to infrastructure/document directory
