+++ host-frontend-root/frontend-src-root/src/infrastructure/document/DomRootChecker.ts
